Separate views from logic using templates
Move the HTML code to a separate `chat.html` file in a `templates` directory.

Implement the `templateHandler` struct using `text/template` and `sync` packages
that is responsible for loading, compiling, and delivering template. The
`templateHandler` type has a single method called `ServeHTTP` that will load the
source file, compile the template and execute it, and write the output to the
specified `http.ResponseWriter` method.